Add vec![None; n] optimization, like vec![0; n]
Similarly to vec![ptr::null{,_mut}(); n] in previous change, this adds
the optimization for vec![None; n] in the case where a zeroed buffer
represents None. A notable exception to this is None for Option<char>,
which, as of 1.24, is represented by 1114112.

There might be convoluted ways to make the optimization work for e.g.
vec![Some('\0'); n], but that doesn't seem worth the effort.

Because this optimization is a little tricky because not all None's are
equal, add a unit test validating that zero-None and non-zero-None are
treated correctly. While here, add tests for the other types.